Use upstream ftgmac100 driver for wedge
Instead of using the hacked MAC driver from aspeed, the change makes the
upstream ftgmac100 driver working on wedge.

One piece of function missing form upstream ftgmac100.c driver, which is
to support the MCAST MAC filter. I ported the code from aspeed driver
for this.

The upstream driver does not have NCSI support, which is one thing left to be
ported over.