Support new rxdes
Each RX descriptor of ftgmac100 has 4 fields (rxdes0, rxdes1, rxdes2,
and rxdes3). Field rxdes2 was not used in early HW, so that the driver
uses it to store the pointer of the page. However, in the new HW used in
AST2500, field rxdes2 is used by HW, which breaks the driver.

The change moved the storage of page pointer from redes2 to a SW field.
The change should work in both early HW and the latest HW.