Fixed up ADC driver
The latest ADC driver only report raw reading, which is useless.
However, in order to do the correct calculation from the reading to the
actual voltage, 3 more parameters (r1, r2, v2) are required.

In the past, we maintain such parameters in the driver. This approach
won't scale as each channel from each board might need different
paramters.

The diff introduced 3 more sysfs to allow user to configure the
parameters. So that, different boards can use different configuration
scripts with the same kernel driver to get the data.

As part of the diff, the old parameters for wedge and Yosemite are
removed. Each board needs to create their own configuration file in
order to use this new driver.