jffs2: kill garbage collect thread when filesystem is corrupted
Reviewed By: zhdaniel12

Differential Revision:
D16987396

fb-note-start
This is a workaround but not official fix: the jffs2 upstream maintainer
was willing to help us but we cannot share the file system image as it
may contain Facebook confidential data. So we will keep the workaround
for now.
The workaround is needed for all BMC platforms depending on jffs2.
fb-note-end

fbshipit-source-id: 18673ff841