jtag: aspeed: add AST2600 support (#45)
Summary:
Enhance the JTAG driver to support the AST2600.

On AST2600, the AST2600-specific JTAG pins are enabled
using an SCU enable pin register.

Also for AST2600, the meaning of the SCU RESET_JTAG
register is different. To set a bit, the SCU_RESET_JTAG
register is used. To clear a bit, an additional clear
register is used.

Reviewed By: joancaneus

Differential Revision: D23093876

Pulled By: joancaneus

fbshipit-source-id: e9f91a52fb4f92b6741fa5ada5c30a8509522baa