fbal: dts: change back to use aspeed_smc driver for SPI1 flash (#1967)
Summary:
Since the BIOS flash and TPM are on separated SPI bus now, it's unnecessary to use
spi-master driver for the BIOS flash.
This also help to align the settings on kernel 5.0 and kernel 5.6 since m25p80
has been removed in kernel 5.6.

Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/1967

Test Plan: Build and test pass on fbal

Reviewed By: tao-ren

Differential Revision: D23871493

Pulled By: amithash

fbshipit-source-id: 3d1dabca67