tpm: add request_locality before write TPM_INT_ENABLE (#1395)
Summary:
Issue Description: TPM Driver Crash causing the device loading to fail.
There was TPM test failure observed on the manufacturer's side. It was determined that the driver crashes during the probe because of which /dev/tpm is not seen. It happens at random, with very rare occurrence.
This issue is seen only on Minipack2 devices.

From the log traces, this issue can be co-related to the root causes similar to those described in: "https://github.com/torvalds/linux/commit/0ef333f5ba7f24f5d8478425c163d3097f1c7afd"

Therefore, we are implementing this preventive fix as suggested by the Linux community. This code will be included in the kernel upgrade and will be a part of Kernel 5.15, so we are porting it back to 5.10. With the preventive fix, we ran 500 iterations without hitting the issue.

Error Logs:
[ErrorLogSnippetTPM.txt](https://github.com/facebookexternal/openbmc.celestica/files/9645980/ErrorLogSnippetTPM.txt)

Test logs with the preventive fix:
[tmp_fix_logs.txt](https://github.com/facebookexternal/openbmc.celestica/files/9644723/tmp_fix_logs.txt)

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1395

Reviewed By: mikechoifb

Differential Revision: D39933393

Pulled By: rootkeeper

fbshipit-source-id: 0fca420bf6f1447a5a51c664641966dc8b56df22