Kernel 5.15: i3c: workaround for IBI PEC error (#2550)
Summary:
Cause:
When an IBI with a repeat start and response of Get CCC is received, the PEC error flag will also be set, cause BMC false alarm PEC Error on dmesg.

Workaround:
When an IBI is received with the PEC error flag set, the patch checks if the received message matches one of the entries in a pre-defined MDB table. If MDB is matched,filter out the warning message. If not matched, will keep the warning.

X-link: https://github.com/facebookexternal/openbmc.wiwynn/pull/2550

Reviewed By: garnermic

Differential Revision: D44169977

Pulled By: williamspatrick

fbshipit-source-id: e64a5e78e25db10d8388858953370d42a2e4c0b4