=?UTF-8?q?[openbmc.quanta][PR]=20kernel5.15:=20ncsi:=20cl?= =?UTF-8?q?ear=20Tx=20enable=20mode=20when=20handling=20Configuration=20re?= =?UTF-8?q?=E2=80=A6?= (#3781)
Summary:
…quired AEN
There's an issue that channel Tx is not enabled when reconfiguring channel triggered by a Configuration required AEN. When reconfiguring the channel, it is misjudged that the channel Tx has been enabled, so that the Enable Channel Network Tx command is not sent. So clear the channel Tx enable flag before reconfiguring the channel to avoid the misjudgment.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/3781

Reviewed By: amithash

Differential Revision: D45083224

fbshipit-source-id: 3ed4ad1f0cad5d253acac5d64ed501ffe5bea566