mtd: spi-nor: winbond: Add support for w25q01jv
Summary:
Add support for Winbond w25q01jv spi-nor chip.

This chip is internally made of two dies with linear addressing capabilities to make it transparent to the user that two dies were used. There is one drawback however, the read status operation is racy as the status bit only gives the active die status and not the status of the other die. For commands affecting the two dies, it means if another command is sent too fast after the first die has returned a valid status (deviation can be up to 200us), the chip will get corrupted/in an unstable state.

This chip hence requires a better status register read. There are three solutions here:
- Either we wait about 200us after getting a first status ready feedback, to cover possible internal deviations.
- Or we always check all internal dies (which takes about 30us per die).

This second option being always faster and also being totally safe, we implement a specific hook for the status register read. flash_speed benchmarks show no difference with this implementation, compared to the regular status read core function, the difference being part of the natural deviation with this benchmark:

	> Without the fixup
	$ flash_speed /dev/mtd0 -c100 -d
	eraseblock write speed is 442 KiB/s
	eraseblock read speed is 1606 KiB/s
	page write speed is 439 KiB/s
	page read speed is 1520 KiB/s
	2 page write speed is 441 KiB/s
	2 page read speed is 1562 KiB/s
	erase speed is 68 KiB/s

	> With the fixup
	$ flash_speed /dev/mtd0 -c100 -d
	eraseblock write speed is 428 KiB/s
	eraseblock read speed is 1626 KiB/s
	page write speed is 426 KiB/s
	page read speed is 1538 KiB/s
	2 page write speed is 426 KiB/s
	2 page read speed is 1574 KiB/s
	erase speed is 66 KiB/s

As there are very few situations where this can actually happen, a status register write being the most likely one, another possibility might have been to use volatile writes instead of non-volatile writes, as most of the deviation comes from the action of writing the bit. But this would overlook other possible actions where both dies can be used at the same time like a chip erase (or any erase over the die boundary in general). This last approach would have the least impact but because it does not feel like it is totally safe to use and because the impact of the second solution presented above is also negligible, we keep this second approach for now (which can be further tuned later if it appears to be too impacting in the end).

However, the fixup, whatever which one we pick, must be applied on multi-die chips, which hence must be properly flagged. The SFDP tables implemented give a lot of information but the die details are part of an optional table that is not implemented, hence we use a post parsing fixup hook to set the params->n_dice value manually.

Link: https://www.winbond.com/resource-files/W25Q01JV%20SPI%20RevE%2003042024%20Plus.pdf
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4622

Reviewed By: amithash

Differential Revision: D72933832

fbshipit-source-id: 8d37dc5d10f11864f914e4ba4a2755b6b4fdb66d